RFCT: rename core to constants for readability and move constants for api keys (query-keys) to global
