+++ frontend/src/features/model-bar/components/UtilityPanel.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-import { navigation_info } from "@/core/navigation";
+import { navigation_info } from "@/constants/navigation";
 import NavSheet from "@/components/widgets/nav-sheet/NavSheet";
 import ModelInstaller from "./ModelInstaller";
 import useModelServices from "../hooks/useModelServices";
